backend/PubSub/HotelManagementTopicGCPFunction/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-pubsub
flask
flask-cors